Zeichenvergleich von Buchstaben muss mittels (unsigned) byte anstatt char erfolgen!

+++ source/ChannelListSaver.c
@@ -449,7 +449,7 @@ bool ConvertUTFStr(char *DestStr, char *SourceStr, int MaxLen, bool ToUnicode)
     if (ToUnicode)
     {
       #ifdef __ALTEFBLIB__
-        if (SourceStr[0] < 0x20) SourceStr++;
+        if ((byte)SourceStr[0] < 0x20) SourceStr++;
         StrToUTF8(SourceStr, TempStr);
       #else
         StrToUTF8(SourceStr, TempStr, 9);
@@ -460,13 +460,13 @@ bool ConvertUTFStr(char *DestStr, char *SourceStr, int MaxLen, bool ToUnicode)
 
     if (!ToUnicode)
     {
-      if ((SourceStr[0] >= 0x20) && (strlen(TempStr) < strlen(SourceStr)))
+      if (((byte)SourceStr[0] >= 0x20) && (strlen(TempStr) < strlen(SourceStr)))
       {
         DestStr[0] = 0x05;
         DestStr++;
         MaxLen--;
       }
-      else if (SourceStr[0] >= 0x15)
+      else if ((byte)SourceStr[0] >= 0x15)
       {
         DestStr[0] = 0x05;
         DestStr++;
